feat: rfox stake/unstake rm status screen
