+++ src/hooks/useActionCenterSubscribers/useGenericTransactionSubscriber.tsx
@@ -1,11 +1,19 @@
 import { fromAccountId } from '@shapeshiftoss/caip'
 import { TxStatus } from '@shapeshiftoss/unchained-client'
+import { useQueryClient } from '@tanstack/react-query'
 import { useEffect } from 'react'
 
 import { useNotificationToast } from '../useNotificationToast'
 
 import { useActionCenterContext } from '@/components/Layout/Header/ActionCenter/ActionCenterContext'
 import { GenericTransactionNotification } from '@/components/Layout/Header/ActionCenter/components/Notifications/GenericTransactionNotification'
+import { getAffiliateRevenueQueryKey } from '@/pages/RFOX/hooks/useAffiliateRevenueQuery'
+import { useCurrentEpochMetadataQuery } from '@/pages/RFOX/hooks/useCurrentEpochMetadataQuery'
+import { getEarnedQueryKey } from '@/pages/RFOX/hooks/useEarnedQuery'
+import { getEpochHistoryQueryKey } from '@/pages/RFOX/hooks/useEpochHistoryQuery'
+import { getStakingBalanceOfQueryKey } from '@/pages/RFOX/hooks/useStakingBalanceOfQuery'
+import { getStakingInfoQueryKey } from '@/pages/RFOX/hooks/useStakingInfoQuery'
+import { getTimeInPoolQueryKey } from '@/pages/RFOX/hooks/useTimeInPoolQuery'
 import { actionSlice } from '@/state/slices/actionSlice/actionSlice'
 import { selectPendingGenericTransactionActions } from '@/state/slices/actionSlice/selectors'
 import {
@@ -24,9 +32,11 @@ export const useGenericTransactionSubscriber = () => {
 
   const pendingGenericTransactionActions = useAppSelector(selectPendingGenericTransactionActions)
   const txs = useAppSelector(selectTxs)
+  const currentEpochMetadataQuery = useCurrentEpochMetadataQuery()
+  const queryClient = useQueryClient()
 
   useEffect(() => {
-    pendingGenericTransactionActions.forEach(action => {
+    pendingGenericTransactionActions.forEach(async action => {
       if (action.status !== ActionStatus.Pending) return
       // RFOX stake/unstake only for now, TODO: handle more
       if (action.transactionMetadata.displayType !== GenericTransactionDisplayType.RFOX) return
@@ -43,6 +53,9 @@ export const useGenericTransactionSubscriber = () => {
       const message =
         action.type === ActionType.Deposit ? 'RFOX.stakeSuccess' : 'RFOX.unstakeSuccess'
 
+      const stakingAssetId = action.transactionMetadata.assetId
+      const stakingAssetAccountId = action.transactionMetadata.accountId
+
       dispatch(
         actionSlice.actions.upsertAction({
           ...action,
@@ -53,6 +66,44 @@ export const useGenericTransactionSubscriber = () => {
           },
         }),
       )
+
+      await queryClient.invalidateQueries({
+        queryKey: getStakingInfoQueryKey({
+          stakingAssetId,
+          stakingAssetAccountId,
+        }),
+      })
+      await queryClient.invalidateQueries({
+        queryKey: getStakingBalanceOfQueryKey({
+          stakingAssetId,
+          accountId: stakingAssetAccountId,
+        }),
+      })
+      await queryClient.invalidateQueries({
+        queryKey: getTimeInPoolQueryKey({
+          stakingAssetId,
+          stakingAssetAccountId,
+        }),
+      })
+      await queryClient.invalidateQueries({
+        queryKey: getEpochHistoryQueryKey(),
+      })
+      await queryClient.invalidateQueries({
+        queryKey: getEarnedQueryKey({
+          stakingAssetId,
+          stakingAssetAccountId,
+        }),
+      })
+      await queryClient.invalidateQueries({
+        queryKey: getAffiliateRevenueQueryKey({
+          startTimestamp: currentEpochMetadataQuery.data?.epochStartTimestamp,
+          endTimestamp: currentEpochMetadataQuery.data?.epochEndTimestamp,
+        }),
+      })
+
+      await queryClient.invalidateQueries({
+        queryKey: ['getUnstakingRequests', { stakingAssetAccountId }],
+      })
 
       // No double-toasty
       if (toast.isActive(action.transactionMetadata.txHash)) return
@@ -79,5 +130,15 @@ export const useGenericTransactionSubscriber = () => {
         },
       })
     })
-  }, [pendingGenericTransactionActions, dispatch, txs, isDrawerOpen, openActionCenter, toast])
+  }, [
+    pendingGenericTransactionActions,
+    dispatch,
+    txs,
+    isDrawerOpen,
+    openActionCenter,
+    toast,
+    currentEpochMetadataQuery.data?.epochEndTimestamp,
+    currentEpochMetadataQuery.data?.epochStartTimestamp,
+    queryClient,
+  ])
 }
